+++ setup/ui.py
@@ -96,7 +96,8 @@ class STORYTOOLS_PT_viewport_setup(bpy.types.Panel):
             col = col.column(align=True)
             row = col.row(align=True)
             row.operator("storytools.create_static_storyboard_pages", text='Create Storyboard Pages', icon='GRID')
-            row.menu("STORYTOOLS_MT_storyboard_presets_management", text="", icon='TOOL_SETTINGS')
+            # row.menu("STORYTOOLS_MT_storyboard_presets_management", text="", icon='TOOL_SETTINGS') # menu for presets management
+            row.menu("STORYTOOLS_MT_static_storyboard_options", text="", icon='TOOL_SETTINGS') # menu for presets management and animatic creation
 
             col.operator("storytools.render_storyboard_images", text='Render StoryBoard', icon='RESTRICT_RENDER_OFF')
             col.menu("STORYTOOLS_MT_export_storyboard_to_pdf", icon='DOCUMENTS', text='Create PDF')
@@ -112,15 +113,73 @@ class STORYTOOLS_PT_viewport_setup(bpy.types.Panel):
             #     if show_storypencil_setup: # Experimental Dual setup
             #         col.operator('storytools.setup_storypencil', text='Setup Storypencil (dual window)', icon='WORKSPACE')
 
+
+## --- Marker management in Timeline
+
+class STORYTOOLS_PT_marker_management(bpy.types.Panel):
+    bl_space_type = "DOPESHEET_EDITOR"
+    # bl_region_type = "HEADER"
+    bl_region_type = "UI"
+    bl_label = "Marker Management"
+    bl_options = {"INSTANCED"}
+
+    def draw(self, context):
+        layout = self.layout
+        layout.label(text="Scale markers:")
+        row = layout.row(align=True)
+        row.operator("storytools.time_compression", text="Compress", icon="TRIA_LEFT").direction = 'COMPRESS'
+        row.operator("storytools.time_compression", text="Dilate", icon="TRIA_RIGHT").direction = 'DILATE'
+
+def marker_management_ui(self, context):
+    """Add a panel to the marker management UI"""
+
+    if not context.scene.storytools_settings.show_marker_management:
+        return
+
+    layout = self.layout
+    layout.label(text="Markers:")
+    row = layout.row(align=True)
+    row.operator("storytools.push_markers", text="", icon="TRIA_LEFT").direction = 'LEFT'
+    row.operator("storytools.push_markers", text="", icon="TRIA_RIGHT").direction = 'RIGHT'
+    row.operator("wm.call_panel", text="", icon='DOWNARROW_HLT').name = "STORYTOOLS_PT_marker_management"
+
+## Options related to static storyboard
+class STORYTOOLS_MT_static_storyboard_options(bpy.types.Menu):
+    bl_label = "Static Storyboard Options"
+
+    def draw(self, context):
+        layout = self.layout
+        layout.menu("STORYTOOLS_MT_storyboard_presets_management", icon='FILE_LARGE', text='Manage Storyboard Presets')
+            
+        layout.label(text='Animatic:')
+        layout.operator("storytools.create_animatic_from_board", text='Create Animatic Scene', icon='SCENE_DATA') # SCENE_DATA, MARKER_HLT ## idea of animatic
+        layout.prop(context.scene.storytools_settings, "show_marker_management", text="Show Marker Management In Timeline")
+        ## TODO: add a quick way to change frame radius size (to avoid getting in animatic frame)
+        # frame_material = bpy.data.materials.get('Frames')
+
+## Options related to static storyboard
 def drawing_setup_ui(self, context):
-    """Drawing Setup pop-up to set in header"""
+    """Drawing Setup pop-up to set in viewport header"""
     self.layout.popover('STORYTOOLS_PT_viewport_setup', text='') # icon='GREASEPENCIL'
 
+
+classes = (
+    STORYTOOLS_PT_marker_management,
+    STORYTOOLS_MT_static_storyboard_options,
+    STORYTOOLS_PT_viewport_setup,
+)
+
 def register():
-    bpy.utils.register_class(STORYTOOLS_PT_viewport_setup)
+    for cls in classes:
+        bpy.utils.register_class(cls)
+
     bpy.types.VIEW3D_HT_header.append(drawing_setup_ui)
+    bpy.types.DOPESHEET_HT_header.append(marker_management_ui)
 
 
 def unregister():
     bpy.types.VIEW3D_HT_header.remove(drawing_setup_ui)
-    bpy.utils.unregister_class(STORYTOOLS_PT_viewport_setup)
+    bpy.types.DOPESHEET_HT_header.remove(marker_management_ui)
+
+    for cls in reversed(classes):
+        bpy.utils.unregister_class(cls)